fix: remove unnecessary composer install step in tests
There is a version mismatch between the composer in the builder image and the one in the site image which causes a fatal error when trying to run  in the site image.
However running  in the tests is not necessary because the dev dependencies are already present in the site image so this command can be removed.

Refs: OPS-7589, OPS-8951